phed datatypes more precise.

+++ sql/phed/root/create_root_phed.sql
@@ -3,10 +3,10 @@ CREATE TABLE public.phed (
     state           VARCHAR(2),
     year            SMALLINT,
     month           SMALLINT,
-    phed_am_peak    REAL,
-    phed_pm_peak_1  REAL,
-    phed_pm_peak_2  REAL,
-    phed_max        REAL
+    phed_am_peak    DOUBLE PRECISION,
+    phed_pm_peak_1  DOUBLE PRECISION,
+    phed_pm_peak_2  DOUBLE PRECISION,
+    phed_max        DOUBLE PRECISION
   )
   WITH (fillfactor=100, autovacuum_enabled=false)
 ;
